feat(auth-rest-passport): Rename from auth-passport
