+++ module/auth-rest-passport/README.md
@@ -1,11 +1,11 @@
 <!-- This file was generated by @travetto/doc and should not be modified directly -->
-<!-- Please modify https://github.com/travetto/travetto/tree/main/module/auth-passport/doc.ts and execute "npx trv doc" to rebuild -->
+<!-- Please modify https://github.com/travetto/travetto/tree/main/module/auth-rest-passport/doc.ts and execute "npx trv doc" to rebuild -->
 # Rest Auth Passport
 ## Rest authentication integration support for the travetto framework
 
-**Install: @travetto/auth-passport**
+**Install: @travetto/auth-rest-passport**
 ```bash
-npm install @travetto/auth-passport
+npm install @travetto/auth-rest-passport
 ```
 
 This is a primary integration for the [Rest Auth](https://github.com/travetto/travetto/tree/main/module/auth-rest#readme "Rest authentication integration support for the travetto framework") module.  This is another level of scaffolding allowing for compatible authentication frameworks to integrate.  
@@ -18,7 +18,7 @@ import { Strategy as FacebookStrategy } from 'passport-facebook';
 
 import { InjectableFactory } from '@travetto/di';
 import { Authenticator, Authorizer, Principal } from '@travetto/auth';
-import { PassportAuthenticator } from '@travetto/auth-passport';
+import { PassportAuthenticator } from '@travetto/auth-rest-passport';
 
 export class FbUser {
   username: string;
@@ -60,7 +60,7 @@ export class AppConfig {
 }
 ```
 
-As you can see, [PassportAuthenticator](https://github.com/travetto/travetto/tree/main/module/auth-passport/src/authenticator.ts#L14) will take care of the majority of the work, and all that is required is:
+As you can see, [PassportAuthenticator](https://github.com/travetto/travetto/tree/main/module/auth-rest-passport/src/authenticator.ts#L14) will take care of the majority of the work, and all that is required is:
    
    *  Provide the name of the strategy (should be unique)
    *  Provide the strategy instance. **Note**: you will need to provide the callback for the strategy to ensure you pass the external principal back into the framework
