+++ module/auth-rest-passport/doc.ts
