+++ module/auth-rest-passport/doc/simple.ts
