+++ module/auth-rest-passport/src/authenticator.ts
