--- related/travetto.github.io/src/app/documentation/gen/auth-passport/auth-passport.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-auth-passport',
-  templateUrl: './auth-passport.component.html',
-  styleUrls: []
-})
-export class AuthPassportComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
-}
