+++ related/travetto.github.io/src/app/documentation/gen/auth-rest-passport/auth-rest-passport.component.html
@@ -1,15 +1,15 @@
 <!-- This file was generated by @travetto/doc and should not be modified directly -->
-<!-- Please modify https://github.com/travetto/travetto/tree/main/module/auth-passport/doc.ts and execute "npx trv doc" to rebuild -->
+<!-- Please modify https://github.com/travetto/travetto/tree/main/module/auth-rest-passport/doc.ts and execute "npx trv doc" to rebuild -->
 <h1>Rest Auth Passport
           <small>Rest authentication integration support for the travetto framework</small>
 
         </h1>
 
       <figure class="install">
-      <figcaption class="install">Install @travetto/auth-passport
+      <figcaption class="install">Install @travetto/auth-rest-passport
       
       </figcaption>
-      <pre><code class="language-bash"><span class="token function">npm</span> <span class="token function">install</span> @travetto/auth-passport</code></pre>     
+      <pre><code class="language-bash"><span class="token function">npm</span> <span class="token function">install</span> @travetto/auth-rest-passport</code></pre>     
       </figure>
 
 This is a primary integration for the <a class="module-link" routerLink="/docs/auth-rest" title="Rest authentication integration support for the travetto framework">Rest Auth</a> module.  This is another level of scaffolding allowing for compatible authentication frameworks to integrate.  
@@ -18,13 +18,13 @@ Within the node ecosystem, the most prevalent auth framework is <a target="_blan
 
       <figure class="code">
       <figcaption class="code">Sample Facebook/passport config
-      <cite><a target="_blank" href="https://github.com/travetto/travetto/tree/main/module/auth-passport/doc/conf.ts">Source</a></cite>
+      <cite><a target="_blank" href="https://github.com/travetto/travetto/tree/main/module/auth-rest-passport/doc/conf.ts">Source</a></cite>
       </figcaption>
       <pre><code class="language-typescript"><span class="token keyword">import</span> <span class="token punctuation">{{'{'}}</span> Strategy <span class="token keyword">as</span> FacebookStrategy <span class="token punctuation">{{'}'}}</span> <span class="token keyword">from</span> <span class="token string">'passport-facebook'</span><span class="token punctuation">;</span>
 
 <span class="token keyword">import</span> <span class="token punctuation">{{'{'}}</span> InjectableFactory <span class="token punctuation">{{'}'}}</span> <span class="token keyword">from</span> <span class="token string">'@travetto/di'</span><span class="token punctuation">;</span>
 <span class="token keyword">import</span> <span class="token punctuation">{{'{'}}</span> Authenticator<span class="token punctuation">,</span> Authorizer<span class="token punctuation">,</span> Principal <span class="token punctuation">{{'}'}}</span> <span class="token keyword">from</span> <span class="token string">'@travetto/auth'</span><span class="token punctuation">;</span>
-<span class="token keyword">import</span> <span class="token punctuation">{{'{'}}</span> PassportAuthenticator <span class="token punctuation">{{'}'}}</span> <span class="token keyword">from</span> <span class="token string">'@travetto/auth-passport'</span><span class="token punctuation">;</span>
+<span class="token keyword">import</span> <span class="token punctuation">{{'{'}}</span> PassportAuthenticator <span class="token punctuation">{{'}'}}</span> <span class="token keyword">from</span> <span class="token string">'@travetto/auth-rest-passport'</span><span class="token punctuation">;</span>
 
 <span class="token keyword">export</span> <span class="token keyword">class</span> <span class="token class-name">FbUser</span> <span class="token punctuation">{{'{'}}</span>
   username<span class="token operator">:</span> <span class="token builtin">string</span><span class="token punctuation">;</span>
@@ -66,7 +66,7 @@ Within the node ecosystem, the most prevalent auth framework is <a target="_blan
 <span class="token punctuation">{{'}'}}</span></code></pre>     
       </figure>
 
-As you can see, <a target="_blank" class="source-link" href="https://github.com/travetto/travetto/tree/main/module/auth-passport/src/authenticator.ts#L14">PassportAuthenticator</a> will take care of the majority of the work, and all that is required is:
+As you can see, <a target="_blank" class="source-link" href="https://github.com/travetto/travetto/tree/main/module/auth-rest-passport/src/authenticator.ts#L14">PassportAuthenticator</a> will take care of the majority of the work, and all that is required is:
 <ul> <li>Provide the name of the strategy (should be unique)</li>
  <li>Provide the strategy instance. <p class="note"><strong>Note</strong> you will need to provide the callback for the strategy to ensure you pass the external principal back into the framework</p></li>
  <li>The conversion functions which defines the mapping between external and local identities.</li></ul>
@@ -75,7 +75,7 @@ After that, the provider is no different than any other, and can be used accordi
 
       <figure class="code">
       <figcaption class="code">Sample routes using Facebook/passport provider
-      <cite><a target="_blank" href="https://github.com/travetto/travetto/tree/main/module/auth-passport/doc/simple.ts">Source</a></cite>
+      <cite><a target="_blank" href="https://github.com/travetto/travetto/tree/main/module/auth-rest-passport/doc/simple.ts">Source</a></cite>
       </figcaption>
       <pre><code class="language-typescript"><span class="token keyword">import</span> <span class="token punctuation">{{'{'}}</span> Controller<span class="token punctuation">,</span> Get<span class="token punctuation">,</span> Redirect<span class="token punctuation">,</span> Post<span class="token punctuation">,</span> Request <span class="token punctuation">{{'}'}}</span> <span class="token keyword">from</span> <span class="token string">'@travetto/rest'</span><span class="token punctuation">;</span>
 <span class="token keyword">import</span> <span class="token punctuation">{{'{'}}</span> AuthService<span class="token punctuation">,</span> Authenticate<span class="token punctuation">,</span> Authenticated<span class="token punctuation">,</span> Unauthenticated <span class="token punctuation">{{'}'}}</span> <span class="token keyword">from</span> <span class="token string">'@travetto/auth-rest'</span><span class="token punctuation">;</span>
