+++ related/travetto.github.io/src/app/documentation/gen/auth-rest-passport/auth-rest-passport.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-auth-rest-passport',
+  templateUrl: './auth-rest-passport.component.html',
+  styleUrls: []
+})
+export class AuthRestPassportComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}
